python_scraper/requirements.txt
selenium
beautifulsoup4
pandas
python-dotenv
supabase
webdriver-manager
PySide6
